polish rich media document details
